+++ src/SqlArtisan/Internal/SqlPart/Clause/ForUpdate/NowaitBehavior.cs
@@ -2,6 +2,8 @@ namespace SqlArtisan.Internal;
 
 public sealed class NowaitBehavior : LockBehaviorBase
 {
+    internal NowaitBehavior() { }
+
     internal override void Format(SqlBuildingBuffer buffer) =>
         buffer.Append(Keywords.Nowait);
 }