+++ src/SqlArtisan/Internal/SqlPart/Condition/ArrayCondition/ArrayContainedByCondition.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class ArrayContainedByCondition(SqlExpression leftSide, SqlExpression rightSide) :
-    ArrayCondition(leftSide, Operators.ArrayContainedBy, rightSide);
+public sealed class ArrayContainedByCondition : ArrayCondition
+{
+    internal ArrayContainedByCondition(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.ArrayContainedBy, rightSide) { }
+}