+++ src/SqlArtisan/Internal/SqlPart/Condition/ArrayCondition/ArrayOverlapsCondition.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class ArrayOverlapsCondition(SqlExpression leftSide, SqlExpression rightSide) :
-    ArrayCondition(leftSide, Operators.ArrayOverlaps, rightSide);
+public sealed class ArrayOverlapsCondition : ArrayCondition
+{
+    internal ArrayOverlapsCondition(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.ArrayOverlaps, rightSide) { }
+}