+++ src/SqlArtisan/Internal/SqlPart/Condition/JsonCondition/JsonbExistsAllCondition.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class JsonbExistsAllCondition(SqlExpression leftSide, SqlExpression[] keys) :
-    JsonbCondition(leftSide, Operators.JsonbExistsAll, new ArrayConstructorExpression(keys));
+public sealed class JsonbExistsAllCondition : JsonbCondition
+{
+    internal JsonbExistsAllCondition(SqlExpression leftSide, SqlExpression[] keys)
+        : base(leftSide, Operators.JsonbExistsAll, new ArrayConstructorExpression(keys)) { }
+}