+++ src/SqlArtisan/Internal/SqlPart/Expression/ArithmeticOperator/DivisionOperator.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class DivisionOperator(SqlExpression leftSide, SqlExpression rightSide) :
-    BinaryOperator(leftSide, Operators.Slash, rightSide);
+public sealed class DivisionOperator : BinaryOperator
+{
+    internal DivisionOperator(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.Slash, rightSide) { }
+}