+++ src/SqlArtisan/Internal/SqlPart/Expression/ArrayConstructorExpression.cs
@@ -1,8 +1,10 @@
 namespace SqlArtisan.Internal;
 
-public sealed class ArrayConstructorExpression(SqlExpression[] elements) : SqlExpression
+public sealed class ArrayConstructorExpression : SqlExpression
 {
-    private readonly SqlExpression[] _elements = elements;
+    private readonly SqlExpression[] _elements;
+
+    internal ArrayConstructorExpression(SqlExpression[] elements) => _elements = elements;
 
     internal override void Format(SqlBuildingBuffer buffer) => buffer
         .Append(Keywords.Array)