+++ src/SqlArtisan/Internal/SqlPart/Expression/JsonOperator/JsonHashArrowTextOperator.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class JsonHashArrowTextOperator(SqlExpression leftSide, SqlExpression rightSide) :
-    BinaryOperator(leftSide, Operators.JsonHashArrowText, rightSide);
+public sealed class JsonHashArrowTextOperator : BinaryOperator
+{
+    internal JsonHashArrowTextOperator(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.JsonHashArrowText, rightSide) { }
+}