+++ src/SqlArtisan/Internal/SqlPart/Expression/VectorOperator/CosineDistanceOperator.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class CosineDistanceOperator(SqlExpression leftSide, SqlExpression rightSide) :
-    BinaryOperator(leftSide, Operators.CosineDistance, rightSide);
+public sealed class CosineDistanceOperator : BinaryOperator
+{
+    internal CosineDistanceOperator(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.CosineDistance, rightSide) { }
+}