+++ src/SqlArtisan/Internal/SqlPart/Expression/VectorOperator/L1DistanceOperator.cs
@@ -1,4 +1,7 @@
 namespace SqlArtisan.Internal;
 
-public sealed class L1DistanceOperator(SqlExpression leftSide, SqlExpression rightSide) :
-    BinaryOperator(leftSide, Operators.L1Distance, rightSide);
+public sealed class L1DistanceOperator : BinaryOperator
+{
+    internal L1DistanceOperator(SqlExpression leftSide, SqlExpression rightSide)
+        : base(leftSide, Operators.L1Distance, rightSide) { }
+}